Add a real tokenizer
This is the first of a series of commits that aim at making a version
of Lust with better error messages and a more modular design. The
added tokenizer emits tokens that have associated [start, end)
locations in the program source. For the moment it assumes that input
is given on one line and just tracks offset from the beginning of the
input although this is scheduled to change.

Of note: at the moment the tokenizer emits an `Eof` token when it
finishes reading the input. This `Eof` token has a location one past
the end of the input string. On one hand I like this because it means
that the `Eof` location doesn't have the same location as the last
token in the input, but on the other hand it is potentially
problematic because indexing the source string at the location of the
`Eof` token has potential to cause out of bounds memory access which
is definitely bad.