Add a parser.
The parser uses a token buffer to eat tokens and returns errors with
locations for rendering as it moves along. At the moment, the parser
is showing an error for inputs like (1 1 1) to the effect of "invalid
list prelude." this is fine for now, but when the type checker comes
along we'll probably want to move that over there.