Add quote special form and skeleton for interpreter.
I think I'm getting ahead of myself here going straight to the
JIT. Seems to me like it would be worthwhile to spend some time on the
language by writing an interpreter and fleshing it out. Macros are a
big part of that plan so hence the addition of the quote syntax sugar.