Stop let from evaluating the value it is binding to.
This is a primarially an ergonomic thing. So far as I can tell the
ergonomic advantage here excedes the practical one by quite a
bit. Later if I add the ability to generate symbols at runtime then it
might be possible to have a `lete` expression or something that has
that power.